// platform/apps/workbench-v2/src/hooks/useTerminalTheme.d.ts
export interface XtermTheme {
    background: string;
    foreground: string;
    cursor: string;
    cursorAccent: string;
    selectionBackground: string;
    black: string;
    red: string;
    green: string;
    yellow: string;
    blue: string;
    magenta: string;
    cyan: string;
    white: string;
    brightBlack: string;
    brightRed: string;
    brightGreen: string;
    brightYellow: string;
    brightBlue: string;
    brightMagenta: string;
    brightCyan: string;
    brightWhite: string;
}
export declare function useTerminalTheme(): XtermTheme;
//# sourceMappingURL=useTerminalTheme.d.ts.map